sync/pyproject.toml
[project]
name = "dt-sync"
version = "0.1.0"
description = "Framework version sync CLI for dynatrace-wwse repos"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
sync = "sync.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["sync/tests"]
